+++ src/journey.js
@@ -198,8 +198,8 @@
         _element_Dialog.style.display = "none";
     }
 
-    function removeFocusClassFromLastElement( fireCustomTrigger ) {
-        fireCustomTrigger = isDefined( fireCustomTrigger ) ? fireCustomTrigger : true;
+    function removeFocusClassFromLastElement( callCustomTrigger ) {
+        callCustomTrigger = isDefined( callCustomTrigger ) ? callCustomTrigger : true;
 
         var bindingOptions = _elements_Attributes_Json[ _elements_Attributes_Keys[ _elements_Attributes_Position ] ];
 
@@ -210,7 +210,7 @@
                 bindingOptions.element.style.position = _element_Focus_Element_PositionStyle;
             }
 
-            if ( fireCustomTrigger ) {
+            if ( callCustomTrigger ) {
                 fireCustomTrigger( bindingOptions.onLeave, bindingOptions.element );
             }
         }
@@ -255,7 +255,7 @@
                     bindingOptions = buildAttributeOptions( bindingOptions.result );
                     bindingOptions.element = element;
 
-                    if ( isDefinedNumber( bindingOptions.order ) ) {
+                    if ( isDefinedNumber( bindingOptions.order ) && ( isDefinedString( bindingOptions.title ) || isDefinedString( bindingOptions.description ) ) ) {
                         _elements_Attributes_Json[ bindingOptions.order ] = bindingOptions;
                         _elements_Attributes_Keys.push( bindingOptions.order );
                     }