+++ docs/CHANGE_LOG.md
@@ -4,7 +4,8 @@
 - Minor documentation improvements.
 - Added a new binding custom trigger called "onClose", which states an event that should be triggered when the dialog is closed on a specific DOM element.
 - Fixed a fault that caused the CSS class "journey-js-element-focus" to remain on an element when the dialog is closed.
-- 
+- A "title" or "description" must be set for the binding options for a DOM element to be accepted.
+- Spacing is now applied for the first item in the dialog (to prevent overlaps with the close button).
 
 <br>